+++ examples/src/main/java/org/apache/spark/streaming/examples/JavaFlumeEventCount.java
@@ -18,6 +18,7 @@
 package org.apache.spark.streaming.examples;
 
 import org.apache.spark.api.java.function.Function;
+import org.apache.spark.examples.streaming.StreamingExamples;
 import org.apache.spark.streaming.*;
 import org.apache.spark.streaming.api.java.*;
 import org.apache.spark.streaming.flume.FlumeUtils;
