+++ examples/src/main/scala/org/apache/spark/examples/streaming/RawNetworkGrep.scala
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.spark.streaming.examples
+package org.apache.spark.examples.streaming
 
 import org.apache.spark.storage.StorageLevel
 import org.apache.spark.streaming._
